+++ models/product.js
@@ -1,0 +1,23 @@
+const mongoose = require('mongoose');
+
+const productSchema = new mongoose.Schema({
+    product_name: {
+      type: String,
+      required: true,
+    },
+    product_price: {
+      type: String,
+      required: true,
+    },
+    isInStock: {
+      type: Boolean,
+      required: true,
+    },
+    category: {
+      type: String,
+      required: true,
+    },
+ });
+
+const ProductModel = mongoose.model("products", productSchema);
+module.exports = ProductModel;
